Add Parser class
