+++ text-ui-test/EXPECTED.TXT
@@ -137,5 +137,60 @@
     3. [E][ ] September  (from: SEP 01 2023 to: SEP 30 2023 )
     ____________________________________________________________
     ____________________________________________________________
+    ☹ OOPS!!! I'm sorry, but I don't know what that means :-(
+    ____________________________________________________________
+    ____________________________________________________________
+    Here are the tasks in your list:
+    1. [T][ ] CS2101 Presentation
+    2. [D][X] CS2103 Individual Project  (by: SEP 01 2023 )
+    3. [E][ ] September  (from: SEP 01 2023 to: SEP 30 2023 )
+    ____________________________________________________________
+    ____________________________________________________________
+    ☹ OOPS!!! I'm sorry, but I don't know what that means :-(
+    ____________________________________________________________
+    ____________________________________________________________
+    Here are the tasks in your list:
+    1. [T][ ] CS2101 Presentation
+    2. [D][X] CS2103 Individual Project  (by: SEP 01 2023 )
+    3. [E][ ] September  (from: SEP 01 2023 to: SEP 30 2023 )
+    ____________________________________________________________
+    ____________________________________________________________
+    ☹ OOPS!!! I'm sorry, but I don't know what that means :-(
+    ____________________________________________________________
+    ____________________________________________________________
+    Here are the tasks in your list:
+    1. [T][ ] CS2101 Presentation
+    2. [D][X] CS2103 Individual Project  (by: SEP 01 2023 )
+    3. [E][ ] September  (from: SEP 01 2023 to: SEP 30 2023 )
+    ____________________________________________________________
+    ____________________________________________________________
+    ☹ OOPS!!! I'm sorry, but I don't know what that means :-(
+    ____________________________________________________________
+    ____________________________________________________________
+    Here are the tasks in your list:
+    1. [T][ ] CS2101 Presentation
+    2. [D][X] CS2103 Individual Project  (by: SEP 01 2023 )
+    3. [E][ ] September  (from: SEP 01 2023 to: SEP 30 2023 )
+    ____________________________________________________________
+    ____________________________________________________________
+    ☹ OOPS!!! I'm sorry, but I don't know what that means :-(
+    ____________________________________________________________
+    ____________________________________________________________
+    Here are the tasks in your list:
+    1. [T][ ] CS2101 Presentation
+    2. [D][X] CS2103 Individual Project  (by: SEP 01 2023 )
+    3. [E][ ] September  (from: SEP 01 2023 to: SEP 30 2023 )
+    ____________________________________________________________
+    ____________________________________________________________
+    Noted. I've removed this task:
+    [T][ ] CS2101 Presentation
+    Now you have 2 tasks in the list.
+    ____________________________________________________________
+    ____________________________________________________________
+    Here are the tasks in your list:
+    1. [D][X] CS2103 Individual Project  (by: SEP 01 2023 )
+    2. [E][ ] September  (from: SEP 01 2023 to: SEP 30 2023 )
+    ____________________________________________________________
+    ____________________________________________________________
     Bye. Hope to see you again soon!
     ____________________________________________________________
